Refactor device restoration test to use setattr/getattr and avoid strict IDE warnings
